requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-filter==25.2
Pillow==12.0.0

